third_party/WebKit/LayoutTests/platform/android/fast/css/image-orientation/image-orientation-expected.txt

Apply image-orientation property and check computed style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS p.style.cssText is "image-orientation: from-image;"
PASS getComputedStyle(p).imageOrientation is "from-image"
PASS p.style.cssText is "image-orientation: 0deg;"
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS p.style.cssText is ""
PASS getComputedStyle(p).imageOrientation is "0deg"
PASS successfullyParsed is true

TEST COMPLETE
